Creating registrar_common and moving registrar to registrar_async
