+++ archive/Makefile
@@ -3,9 +3,9 @@ include $(ROOT)/functions.Makefile
 
 STACK_ROOT 	= archive
 
-SBT_APPS    = archivist registrar progress_async progress_http notifier registar_http
+SBT_APPS    = archivist progress_async progress_http notifier registrar_async registrar_http
 SBT_DOCKER_LIBRARIES = archive_common
-SBT_NO_DOCKER_LIBRARIES =
+SBT_NO_DOCKER_LIBRARIES = registrar_common
 
 ECS_TASKS   = bagger archive_api callback_stub_server
 LAMBDAS     =
