--- archive/registrar/src/main/scala/uk/ac/wellcome/platform/archive/registrar/models/RegistrationCompleteNotification.scala
@@ -1,8 +1,0 @@
-package uk.ac.wellcome.platform.archive.registrar.models
-
-import java.util.UUID
-
-import uk.ac.wellcome.platform.archive.common.models.BagId
-
-case class RegistrationCompleteNotification(archiveRequestId: UUID,
-                                            bagId: BagId)
