+++ archive/registrar_async/src/main/scala/uk/ac/wellcome/platform/archive/registrar/async/models/RegistrationCompleteNotification.scala
@@ -1,0 +1,7 @@
+package uk.ac.wellcome.platform.archive.registrar.async.models
+import java.util.UUID
+
+import uk.ac.wellcome.platform.archive.common.models.BagId
+
+case class RegistrationCompleteNotification(archiveRequestId: UUID,
+                                            bagId: BagId)
