+++ archive/registrar_async/src/test/resources/application.conf
