+++ archive/registrar_async/src/test/resources/logback.xml
